added checks for invalid user input
