+++ src/shortened_url.py
@@ -1,0 +1,7 @@
+def is_shortened_url(shortened_url):
+    """
+    Checks that the requested shortened url exists. CURRENTLY USING PLACEHOLDER VALUE.
+    :param shortened_url: A string to check for if its a valid shortened url.
+    :return: boolean value for if the shortened url is valid.
+    """
+    return shortened_url == "valid_url"
